+++ src/dune_package.mli
@@ -5,7 +5,7 @@ module Lib : sig
 
   val dir : _ t -> Path.t
   val orig_src_dir : _ t -> Path.t option
-  val obj_dir : _ t -> Obj_dir.t
+  val obj_dir : _ t -> Path.t Obj_dir.t
   val requires : _ t -> (Loc.t * Lib_name.t) list
   val name : _ t -> Lib_name.t
   val version : _ t -> string option
@@ -23,7 +23,7 @@ module Lib : sig
   val plugins : _ t -> Path.t list Mode.Dict.t
   val jsoo_runtime : _ t -> Path.t list
   val implements : _ t -> (Loc.t * Lib_name.t) option
-  val variant : _ t -> Variant.t option
+  val known_implementations : _ t -> (Loc.t * Lib_name.t) Variant.Map.t
   val default_implementation : _ t -> (Loc.t * Lib_name.t) option
   val special_builtin_support
     : _ t -> Dune_file.Library.Special_builtin_support.t option
@@ -51,14 +51,14 @@ module Lib : sig
     -> requires:(Loc.t * Lib_name.t) list
     -> ppx_runtime_deps:(Loc.t * Lib_name.t) list
     -> implements:(Loc.t * Lib_name.t) option
-    -> variant: (Variant.t) option
     -> default_implementation: (Loc.t * Lib_name.t) option
     -> virtual_:bool
+    -> known_implementations: (Loc.t * Lib_name.t) Variant.Map.t
     -> modules:Lib_modules.t option
     -> modes:Mode.Dict.Set.t
     -> version:string option
     -> orig_src_dir:Path.t option
-    -> obj_dir:Obj_dir.t
+    -> obj_dir:Path.t Obj_dir.t
     -> special_builtin_support:
          Dune_file.Library.Special_builtin_support.t option
     -> 'a t