+++ test/blackbox-tests/test-cases/variants-external-declaration/prj2/vlibfoo.ml
@@ -1,0 +1,1 @@
+let implme () = print_endline "foobar"